RAG/RAG_Debutant/requirements.txt
# ============================================================
# RAG Project — Version LangChain (100% open-source)
# ============================================================

# --- LangChain core ---
langchain==0.2.16
langchain-community==0.2.17
langchain-huggingface==0.0.6
langchain-chroma==0.1.4          # Intégration ChromaDB officielle

# --- Embeddings (open-source, local) ---
sentence-transformers==3.0.1
torch>=2.0.0

# --- Vector Store ---
chromadb==0.5.15

# --- LLM local ---
ollama==0.3.3

# --- LLM HuggingFace ---
huggingface-hub==0.24.6
transformers==4.44.2

# --- Document loaders ---
pypdf==4.3.1
python-docx==1.1.2

# --- Sources open-source ---
wikipedia-api==0.6.0
arxiv==2.1.3
datasets==2.21.0

# --- Utilitaires ---
python-dotenv==1.0.1
tqdm==4.66.5
rich==13.8.0
pydantic==2.8.2

# --- Évaluation RAG (RAGAS) ---
ragas==0.1.21                    # Framework d'évaluation RAG open-source

# --- Tests ---
pytest==8.3.2
pytest-asyncio==0.23.8           # Tests async (utilisé par RAGAS en interne)